+++ tests/Rfc4122/UuidV1Test.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Ramsey\Uuid\Test\Rfc4122;
 
+use DateTimeImmutable;
 use Mockery;
 use Ramsey\Uuid\Codec\CodecInterface;
 use Ramsey\Uuid\Converter\NumberConverterInterface;
@@ -12,6 +13,7 @@ use Ramsey\Uuid\Exception\InvalidArgumentException;
 use Ramsey\Uuid\Rfc4122\FieldsInterface;
 use Ramsey\Uuid\Rfc4122\UuidV1;
 use Ramsey\Uuid\Test\TestCase;
+use Ramsey\Uuid\Uuid;
 
 class UuidV1Test extends TestCase
 {
@@ -52,6 +54,45 @@ class UuidV1Test extends TestCase
             ['version' => 7],
             ['version' => 8],
             ['version' => 9],
+        ];
+    }
+
+    /**
+     * @dataProvider provideUuidV1WithOddMicroseconds
+     */
+    public function testGetDateTimeProperlyHandlesLongMicroseconds(string $uuid, string $expected): void
+    {
+        /** @var UuidV1 $object */
+        $object = Uuid::fromString($uuid);
+
+        $date = $object->getDateTime();
+
+        $this->assertInstanceOf(DateTimeImmutable::class, $date);
+        $this->assertSame($expected, $date->format('U.u'));
+    }
+
+    /**
+     * @phpcsSuppress SlevomatCodingStandard.TypeHints.ReturnTypeHint.MissingTraversableTypeHintSpecification
+     */
+    public function provideUuidV1WithOddMicroseconds(): array
+    {
+        return [
+            [
+                'uuid' => '14814000-1dd2-11b2-9669-00007ffffffe',
+                'expected' => '1.677722',
+            ],
+            [
+                'uuid' => '13714000-1dd2-11b2-9669-00007ffffffe',
+                'expected' => '0.104858',
+            ],
+            [
+                'uuid' => '13713000-1dd2-11b2-9669-00007ffffffe',
+                'expected' => '0.105267',
+            ],
+            [
+                'uuid' => '12e8a980-1dd2-11b2-8d4f-acde48001122',
+                'expected' => '-1.000000',
+            ],
         ];
     }
 }